+++ app/Actions/Video/ConvertVideoAction.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
+namespace Modules\Media\Actions\Video;
+
 /**
  * @see https://github.com/protonemedia/laravel-ffmpeg
  */
-
-namespace Modules\Media\Actions\Video;
-
 use FFMpeg\Format\Video\X264;
 use Illuminate\Support\Facades\Storage;
+use Modules\Media\Support\Ffmpeg\MediaExporterResolver;
 use ProtoneMedia\LaravelFFMpeg\Support\FFMpeg;
 use Spatie\QueueableAction\QueueableAction;
 
@@ -28,12 +28,17 @@ class ConvertVideoAction
 
         $exportedMedia = $openedMedia->export();
 
-        $format = new X264();
+        $format = new X264;
         $format->setKiloBitrate(1000);
 
         $exportedMedia->toDisk($disk_mp4);
         $exportedMedia->inFormat($format);
         $exportedMedia->save($file_new);
+        $format = new X264;
+        $formattedMedia = MediaExporterResolver::from(
+            $exportedMedia->toDisk($disk_mp4)
+        )->inFormat($format);
+        $formattedMedia->save($file_new);
 
         return Storage::disk($disk_mp4)->url($file_new);
     }