+++ app/Console/Commands/ConvertVideoCommand.php
@@ -8,6 +8,7 @@ use FFMpeg\Format\Video\WebM;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
+use Modules\Media\Support\Ffmpeg\MediaExporterResolver;
 use ProtoneMedia\LaravelFFMpeg\Support\FFMpeg;
 use Webmozart\Assert\Assert;
 
@@ -30,7 +31,8 @@ class ConvertVideoCommand extends Command
             return '';
         }
 
-        $format = new WebM();
+        $format = new WebM;
+        $format = new WebM;
         $extension = mb_strtolower(class_basename($format));
         $file_new = Str::of($file)->replaceLast('.mp4', '.'.$extension)->toString();
 
@@ -45,6 +47,10 @@ class ConvertVideoCommand extends Command
         $export->toDisk($disk);
         $export->inFormat($format);
         $export->save($file_new);
+        $formattedExport = MediaExporterResolver::from(
+            $export->toDisk($disk)
+        )->inFormat($format);
+        $formattedExport->save($file_new);
 
         return Storage::disk($disk)->url($file_new);
     }