+++ app/Filament/Resources/MediaResource/Schemas/MediaForm.php
@@ -5,12 +5,13 @@ declare(strict_types=1);
 namespace Modules\Media\Filament\Resources\MediaResource\Schemas;
 
 use Filament\Forms\Components\TextInput;
+use Filament\Schemas\Components\Component;
 use Modules\Xot\Filament\Resources\Schemas\XotBaseResourceForm;
 
 class MediaForm extends XotBaseResourceForm
 {
     /**
-     * @return array<string, \Filament\Schemas\Components\Component>
+     * @return array<string, Component>
      */
     public static function getFormSchema(): array
     {