+++ app/Filament/Resources/TemporaryUploadResource/Tables/TemporaryUploadsTable.php
@@ -4,13 +4,21 @@ declare(strict_types=1);
 
 namespace Modules\Media\Filament\Resources\TemporaryUploadResource\Tables;
 
+use Filament\Actions\DeleteAction;
+use Filament\Actions\DeleteBulkAction;
+use Filament\Actions\EditAction;
+use Filament\Actions\ViewAction;
+use Filament\Tables\Columns\Column;
 use Filament\Tables\Columns\TextColumn;
+use Filament\Tables\Filters\SelectFilter;
+use Modules\Media\Models\TemporaryUpload;
 use Modules\Xot\Filament\Resources\Tables\XotBaseResourceTable;
 
 class TemporaryUploadsTable extends XotBaseResourceTable
 {
     /**
-     * @return array<string, \Filament\Tables\Columns\Column>
+     * @return array<string, Column>
+     * @return array<string, Column>
      */
     public function getTableColumns(): array
     {
@@ -18,6 +26,41 @@ class TemporaryUploadsTable extends XotBaseResourceTable
             'id' => TextColumn::make('id')->searchable()->sortable(),
             'created_at' => TextColumn::make('created_at')->dateTime(),
             'updated_at' => TextColumn::make('updated_at')->dateTime(),
+        ];
+    }
+
+    /**
+     * @return array<string, SelectFilter>
+     */
+    public function getTableFilters(): array
+    {
+        return [
+            'folder' => SelectFilter::make('folder')->options(TemporaryUpload::distinct()->pluck(
+                'folder',
+                'folder',
+            )->toArray(...)),
+        ];
+    }
+
+    /**
+     * @return array<string, ViewAction|EditAction|DeleteAction>
+     */
+    public function getTableActions(): array
+    {
+        return [
+            'view' => ViewAction::make(),
+            'edit' => EditAction::make(),
+            'delete' => DeleteAction::make(),
+        ];
+    }
+
+    /**
+     * @return array<string, DeleteBulkAction>
+     */
+    public function getTableBulkActions(): array
+    {
+        return [
+            'delete' => DeleteBulkAction::make(),
         ];
     }
 }