+++ app/Models/MediaConvert.php
@@ -39,7 +39,9 @@ use Modules\Xot\Contracts\ProfileContract;
  * @property string|null $converted_file
  * @property string|null $disk
  * @property string|null $file
+ * @property string|null $path
  * @property Media|null $media
+ *
  * @method static MediaConvertFactory factory($count = null, $state = [])
  * @method static Builder|MediaConvert newModelQuery()
  * @method static Builder|MediaConvert newQuery()
@@ -65,9 +67,11 @@ use Modules\Xot\Contracts\ProfileContract;
  * @method static Builder|MediaConvert whereUpdatedAt($value)
  * @method static Builder|MediaConvert whereUpdatedBy($value)
  * @method static Builder|MediaConvert whereWidth($value)
+ *
  * @property-read ProfileContract|null $creator
  * @property-read ProfileContract|null $updater
  * @property-read ProfileContract|null $deleter
+ *
  * @mixin \Eloquent
  */
 class MediaConvert extends BaseModel