+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.0.3-dev.12](https://github.com/laraxot/module_media_fila5/compare/v0.0.3-dev.11...v0.0.3-dev.12) (2026-08-24)
+
+### Bug Fixes
+
+* **media:** make the temporary upload path generator work with the real key type ([15b686a](https://github.com/laraxot/module_media_fila5/commit/15b686a9f526f60e98d52dcf0ee0b72e3bc9d45c))
+* **phpstan:** type closure and method params to close typeCoverage.paramTypeCoverage ([eb7dac9](https://github.com/laraxot/module_media_fila5/commit/eb7dac9ac908d15708eaa4b499ce1a3cc10f08c9))
+
 ## [0.0.3-dev.2](https://github.com/laraxot/module_media_fila5/compare/v0.0.3-dev.1...v0.0.3-dev.2) (2026-07-08)
 
 ### Bug Fixes